requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pandas>=2.2.0
python-dotenv>=1.0.0
tqdm>=4.66.0
sentence-transformers>=3.0.0
scikit-learn>=1.4.0
numpy>=1.26.0
